Moved test util that is only used in one file into the file
